Add support for GPG keys (#19)
The upload command now takes care of generating Release files and optionally signing them as InRelease files using a GPG key. This is enabled with the --sign flag. The GPG key can be provided from the keyring using the --gpg-key-id flag, which is a good fit for developer machines. Alternatively, for CI, the GPG key can be provided as an environment variable, $GPG_SIGNING_KEY. This key will be added to the keyring and used. If the GPG key has a password, it can be provided with the $GPG_PASSWORD environment variable.

Fixes #13